requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
python-dotenv
google-api-python-client
google-auth-oauthlib
